Create Hospital.cpp
A program to create a hospital database regarding its indoor patients. The information to store includes: 
a) name of the patient 
b) date of admission 
c) disease 
d) date of discharge
Create a base class to store the above information. The member function should include functions to enter information and display a list of all the patients in the database. Create a derived class to store the age of patients. List the information about all the pediatric patients less than 12 years of age.